introduce server config, refactor
